fix incorrect to_id in nearest()
I believe this fixes issue #328 . `to_int <- as.data.frame(intersect(to, y))` results in a data frame that is out of order with respect to the input data frame `to`, causing incorrect `to_id` in the output of `nearest()`. I suspect there may be a better approach here, mostly submitting this PR to point to where the issue is occurring.